add Japanese dictionary and punctuation
The original Japanese model only used a long list of Kanji which proved
to be suboptimal to recognise particular Japanese constructs.

I added a script that starting from JMdict and JMnedict, creates a wordlist
using Hiragana, Katakana and Kanji variants. It then creates a character
file from this wordlist.

I also added a punctuation file with the characters that made sense from
the Japanese punctuation article on Wikipedia [1].

[1]: https://en.wikipedia.org/wiki/List_of_Japanese_typographic_symbols